auth_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pypsrp==0.8.1
pyjwt==2.10.1
python-dotenv==1.0.1
pydantic==2.10.4 